+++ sacrerouge/__init__.py
@@ -1,2 +1,2 @@
-from sacrerouge.__main__ import build_argument_parser
+from sacrerouge.arguments import build_argument_parser
 from sacrerouge.version import VERSION as __version__
